move eth vault user utils to e2e utils
